Tests/LibWeb/Text/expected/css/computed-style-box-sizing.txt

border-box width: 120px
border-box height: 80px
content-box width: 120px
content-box height: 80px
